feat(Time): 관리자 예약 시간 삭제 API 예외 처리 추가

+++ src/main/java/roomescape/domain/time/service/TimeService.java
@@ -61,6 +61,10 @@ public class TimeService {
     }
 
     public void deleteTimeById(Long id) {
+        if (!timeRepository.existsTimeById(id)) {
+            throw new GeneralException(TimeErrorType.TIME_NOT_FOUND);
+        }
+
         timeRepository.deleteTimeById(id);
     }
 }
